Stop displaying ops:packaged
At the Austin Summit, working with package maintainers, the ops
tags team decided on a way to markedly improve the ops:packaged tag:

https://etherpad.openstack.org/p/AUS-ops-Ops-tags

Unfortunately, these changes have not been implemented and the tag
has not been updated as a result.

Until these changes have been implemented, the ops:package tag should
not be displayed on the project navigator.